Upgrade group samples
