+++ 2-advanced/dubbo-samples-group/dubbo-samples-group-interface/src/main/java/org/apache/dubbo/samples/group/api/GroupService.java
